Export of internal ZetaSQL changes.
  - Add extractTableNamesFromScript to java by Matthew Brown <matthewbr@google.com>
  - Add RENAME TO action to ALTER TABLE. by ZetaSQL Team <no-reply@google.com>
  - Update lambda argument list error msg. by ZetaSQL Team <no-reply@google.com>
  - Add basic lambda support to FunctionArgumentType. by ZetaSQL Team <no-reply@google.com>
  - Avoid crashing for recursive protos in HasSubfieldWithFor... by ZetaSQL Team <no-reply@google.com>
  - Add basic lambda support to FunctionArgumentType. by ZetaSQL Team <no-reply@google.com>
  - REGEXP_INSTR, Code the reference implementation and add c... by ZetaSQL Team <no-reply@google.com>
  - Disable has_ for map entries. by ZetaSQL Team <no-reply@google.com>
  - local_service stores AnalyzerOptions by Andrew Pilloud <apilloud@google.com>
  - Rename ASTLambda::parameter_list to argument_list by ZetaSQL Team <no-reply@google.com>
  - Add an ASTDdlStatement class. by ZetaSQL Team <no-reply@google.com>
  - Add function(s) ProtoType::HasSubfieldWithFormat() to che... by ZetaSQL Team <no-reply@google.com>
  - Extract CollectTemplatedArguments For simple types for ea... by ZetaSQL Team <no-reply@google.com>
  - Add OCTET_LENGTH alias for BYTE_LENGTH by ZetaSQL Team <no-reply@google.com>
  - Fix the explanation of the cumulative sum example. by ZetaSQL Team <no-reply@google.com>
  - Fixes to compliance tests and reference implementation fo... by ZetaSQL Team <no-reply@google.com>
  - Fix the confusing error message: by ZetaSQL Team <no-reply@google.com>
  - Add GEOGRAPHY and JSON data types to unsupported argument... by ZetaSQL Team <no-reply@google.com>
  - Avoid divide by zero with doubles. by ZetaSQL Team <no-reply@google.com>
  - Fix an integer overflow in ROUND and TRUNC implementation... by ZetaSQL Team <no-reply@google.com>
  - Adds link to type definitions with min/max ranges. by ZetaSQL Team <no-reply@google.com>
  - Support converting an encoded decimal value with arbitrar... by ZetaSQL Team <no-reply@google.com>
  - Fix scripting tests to use the PreparedModify api for DML... by ZetaSQL Team <no-reply@google.com>

GitOrigin-RevId: 279bd4c52068b423117f3e745ba70ecfaf2bacb8

+++ zetasql/common/multiprecision_int_impl.h
@@ -592,7 +592,7 @@ int LongDiv(std::array<uint32_t, size + 1>* dividend,
   // Art of Computer Programming" Vol.2 for details.
   //
   // We perform normalization by finding how far we need to shift to make the
-  // most significant bit of the divisor 1. And then we shift both the divident
+  // most significant bit of the divisor 1. And then we shift both the dividend
   // and the divisor thus preserving the result of the division.
   int non_zero_bit_idx = FindMSBSetNonZero((*divisor)[n - 1]);
   int shift_amount = 32 - non_zero_bit_idx - 1;
@@ -607,7 +607,7 @@ int LongDiv(std::array<uint32_t, size + 1>* dividend,
   for (int i = m - n; i >= 0; --i, --dividend_data) {
     // Make the guess of the quotient digit. The guess we take here is:
     //
-    //   qhat = min((divident[m] * b + divident[m-1]) / divisor[n], b-1)
+    //   qhat = min((dividend[m] * b + dividend[m-1]) / divisor[n], b-1)
     //
     // where b is the radix, which in our case is 2^32. In "The Art
     // of Computer Programming" Vol.2 Knuth proves that given the normalization
@@ -763,7 +763,7 @@ bool ParseFromBase10UnsignedString(absl::string_view str, Word* result) {
 
 // Appends 9-digit segments to result in decimal format. Each segment must be
 // <= 999999999. They must be in little endian order.
-void AppendSegmentsToString(const uint32_t segments[], int num_segments,
+void AppendSegmentsToString(const uint32_t segments[], size_t num_segments,
                             std::string* result);
 
 // The following functions are not optimized for performance, but they can be
